task 7 typecasted sqrt of size

+++ 0x1E-search_algorithms/100-jump.c
@@ -14,7 +14,7 @@ int jump_search(int *array, size_t size, int value)
 
 	if (array == NULL)
 		return (-1);
-	step = sqrt(size);
+	step = (int)sqrt(size);
 	for (i = 0; i < (int)size && array[i] < value; i += step)
 	{
 		printf("Value checked array[%d] = [%d]\n", i, array[i]);
